+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [Header] Make rune price visible for smaller window sizes [#1880](https://github.com/thorchain/asgardex-electron/pull/1880)
 - [Header] Tooltip to explain VOLUME (24h) [#1885](https://github.com/thorchain/asgardex-electron/pull/1885)
 - [ERC] Update whitelist (incl. icon support for ETH.THOR, ETH.FOX) [#1894](https://github.com/thorchain/asgardex-electron/issues/1894), [#1896](https://github.com/thorchain/asgardex-electron/issues/1896)
+- Use sync mode for broadcasting transactions to THORChain [x1919](https://github.com/thorchain/asgardex-electron/pull/1919)
 
 ## Fix